+++ src/main/java/org/ohdsi/webapi/security/authc/LoginController.java
@@ -1,6 +1,8 @@
 package org.ohdsi.webapi.security.authc;
 
 import java.util.List;
+import java.util.Optional;
+import java.util.UUID;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -214,5 +216,42 @@ public class LoginController {
       return loginSvc.onSuccess(authenticatedLogin);
     }
   }
+
+  public static class OTC {
+
+    private final OneTimeCodeService oneTimeCodeService;
+    private static final Logger log = LoggerFactory.getLogger(OTC.class);
+
+    public OTC(OneTimeCodeService oneTimeCodeService) {
+      this.oneTimeCodeService = oneTimeCodeService;
+    }
+
+    /**
+     * Redeem a one-time code for a JWT token.
+     * 
+     * @param code the OTC UUID to redeem
+     * @return JWT token wrapped in LoginService.Result, or 401 if OTC is invalid/expired/revoked
+     */
+    @GetMapping(value = "/otc", produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<LoginService.Result> redeemOtc(@RequestParam UUID code) {
+
+      Optional<OneTimeCodeEntity> otcEntity = oneTimeCodeService.validateAndConsume(code);
+
+      if (otcEntity.isEmpty()) {
+        log.debug("OTC redemption failed: code {} invalid or expired", code);
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+            .body(new LoginService.Result(null, null, null, "Invalid or expired code"));
+      }
+
+      OneTimeCodeEntity otc = otcEntity.get();
+      String jwt = otc.getJwtToken();
+
+      log.debug("OTC: {} redeemed successfully for user: {}", code, otc.getLogin());
+
+      // Return the pre-minted JWT directly
+      // Client uses this token to establish session with WebAPI
+      return ResponseEntity.ok(new LoginService.Result(otc.getLogin(), jwt, null, "OTC redeemed successfully."));
+    }
+  }  
 
 }
