+++ src/main/java/org/ohdsi/webapi/security/authc/OneTimeCodeService.java
@@ -22,6 +22,7 @@ import java.util.UUID;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -114,6 +115,7 @@ public class OneTimeCodeService {
    * Clean up expired and revoked one-time codes.
    * Called periodically to remove old OTC records.
    */
+  @Scheduled(fixedDelayString = "#{@oneTimeCodeProperties.ttl.toMillis()}")
   public void cleanupExpiredCodes() {
     Instant now = Instant.now();
     repo.deleteByExpiresAtBefore(now);
